Added MPI_LONG to broadcast call for numActiveKernels array
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@9618 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/connections/HyPerConn.cpp
@@ -3187,7 +3187,7 @@ void HyPerConn::reduceNumKernelActivations(){
       //Do mpi to update numKernelActivationss
       for (int arbor = 0; arbor < numAxonalArborLists; arbor++) {
          int numKernelIndices = getNumDataPatches();
-         int ierr = MPI_Allreduce(MPI_IN_PLACE, numKernelActivations[arbor][0], numKernelIndices*nxp*nyp*nfp, MPI_INT, MPI_SUM, parent->icCommunicator()->communicator());
+         int ierr = MPI_Allreduce(MPI_IN_PLACE, numKernelActivations[arbor][0], numKernelIndices*nxp*nyp*nfp, MPI_LONG, MPI_SUM, parent->icCommunicator()->communicator());
       }
 #endif
    }
